+++ src/file/profilePicture/entities/profile-picture-type.enum.ts
